Refactor private note endpoint
