+++ packages/nextjs/app/api/large-stages/[stageId]/private-note/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest } from "next/server";
-import { createPrivateNoteRequest } from "~~/app/api/stages/[stageId]/private-note/route";
+import { createPrivateNoteRequest } from "~~/utils/privateNote";
 
 export async function POST(req: NextRequest, { params }: { params: { stageId: string } }) {
   return createPrivateNoteRequest(req, { params, isLargeGrant: true });
